fix: redirect user on invalid journey
