+++ apps/journeys-admin/src/libs/useInvalidJourneyRedirect/useInvalidJourneyRedirect.ts
@@ -1,0 +1,12 @@
+import { useEffect } from 'react'
+import { useRouter } from 'next/router'
+
+export const useInvalidJourneyRedirect = (data): void => {
+  const router = useRouter()
+  useEffect(() => {
+    const checkValidJourney = (): void => {
+      data?.journey == null && data?.journey !== undefined && router.push('/')
+    }
+    checkValidJourney()
+  }, [data, router])
+}
